+++ core-ng/src/main/java/core/framework/impl/web/HTTPServerIOHandler.java
@@ -37,7 +37,7 @@ public class HTTPServerIOHandler implements HttpHandler {
             TextBodyReader reader = new TextBodyReader(exchange, handler);
             StreamSourceChannel channel = exchange.getRequestChannel();
             reader.read(channel);  // channel will be null if getRequestChannel() is already called, but here should not be that case
-            if (!reader.finished()) {
+            if (!reader.complete()) {
                 channel.getReadSetter().set(reader);
                 channel.resumeReads();
                 return;