+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Change log
 
+### 3.6.2 (9/25/2015 - 9/28/2015)
+* util: added ByteBuf for high performance network IO and NIO
+* http-client: response bytes() changed to inputStream()
+
 ### 3.6.1 (9/21/2015 - 9/25/2015)
 * web-service: moved core.framework.api.web.client.WebServiceRequestSigner to core.framework.api.web.service.WebServiceRequestSigner
 * web: changed form parsing threading model, support file upload